requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3
python-dotenv==1.0.1
duckduckgo-search==7.2.1
requests==2.32.3
